pyproject.toml
[project]
name = "gw2-data"
version = "0.1.0"
description = "Static acquisition dataset for Guild Wars 2 legendary crafting"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "pydantic>=2.5",
    "httpx>=0.27",
    "anthropic>=0.40",
    "diskcache>=5.6.3",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
